Move HomeListViewControllerTests to unit tests group
